docs(config): clarify config.json
